component registration reworked
