+++ src/physics/physics_system.cpp
@@ -16,161 +16,6 @@
 
 namespace Lumix
 {
-	static void registerProperties(IAllocator& allocator)
-	{
-		using namespace Reflection;
-
-		struct LayerEnum : Reflection::EnumAttribute {
-			u32 count(ComponentUID cmp) const override { 
-				return ((PhysicsScene*)cmp.scene)->getSystem().getCollisionsLayersCount();
-			}
-			const char* name(ComponentUID cmp, u32 idx) const override { 
-				PhysicsSystem& system = ((PhysicsScene*)cmp.scene)->getSystem();
-				return system.getCollisionLayerName(idx);
-			}
-		};
-
-		struct DynamicTypeEnum : Reflection::EnumAttribute {
-			u32 count(ComponentUID cmp) const override { return 3; }
-			const char* name(ComponentUID cmp, u32 idx) const override { 
-				switch ((PhysicsScene::DynamicType)idx) {
-					case PhysicsScene::DynamicType::DYNAMIC: return "Dynamic";
-					case PhysicsScene::DynamicType::STATIC: return "Static";
-					case PhysicsScene::DynamicType::KINEMATIC: return "Kinematic";
-					default: ASSERT(false); return "N/A";
-				}
-			}
-		};
-
-		struct D6MotionEnum : Reflection::EnumAttribute {
-			u32 count(ComponentUID cmp) const override { return 3; }
-			const char* name(ComponentUID cmp, u32 idx) const override { 
-				switch ((PhysicsScene::D6Motion)idx) {
-					case PhysicsScene::D6Motion::LOCKED: return "Locked";
-					case PhysicsScene::D6Motion::LIMITED: return "Limited";
-					case PhysicsScene::D6Motion::FREE: return "Free";
-					default: ASSERT(false); return "N/A";
-				}
-			}
-		};
-
-		struct WheelSlotEnum : Reflection::EnumAttribute {
-			u32 count(ComponentUID cmp) const override { return 4; }
-			const char* name(ComponentUID cmp, u32 idx) const override { 
-				switch ((PhysicsScene::WheelSlot)idx) {
-					case PhysicsScene::WheelSlot::FRONT_LEFT: return "Front left";
-					case PhysicsScene::WheelSlot::FRONT_RIGHT: return "Front right";
-					case PhysicsScene::WheelSlot::REAR_LEFT: return "Rear left";
-					case PhysicsScene::WheelSlot::REAR_RIGHT: return "Rear right";
-					default: ASSERT(false); return "N/A";
-				}
-			}
-		};
-
-		static auto phy_scene = scene("physics",
-			functions(
-				LUMIX_FUNC(PhysicsScene::raycast)
-				//function(LUMIX_FUNC(PhysicsScene::raycastEx))
-			),
-			component("ragdoll",
-				functions(
-					LUMIX_FUNC(PhysicsScene::setRagdollKinematic)
-				),
-				blob_property("data", LUMIX_PROP(PhysicsScene, RagdollData)),
-				property("Layer", LUMIX_PROP(PhysicsScene, RagdollLayer), LayerEnum())
-			),
-			component("d6_joint",
-				property("Connected body", LUMIX_PROP(PhysicsScene, JointConnectedBody)),
-				property("Axis position", LUMIX_PROP(PhysicsScene, JointAxisPosition)),
-				property("Axis direction", LUMIX_PROP(PhysicsScene, JointAxisDirection)),
-				enum_property("X motion", LUMIX_PROP(PhysicsScene, D6JointXMotion), D6MotionEnum()),
-				enum_property("Y motion", LUMIX_PROP(PhysicsScene, D6JointYMotion), D6MotionEnum()),
-				enum_property("Z motion", LUMIX_PROP(PhysicsScene, D6JointZMotion), D6MotionEnum()),
-				enum_property("Swing 1", LUMIX_PROP(PhysicsScene, D6JointSwing1Motion), D6MotionEnum()),
-				enum_property("Swing 2", LUMIX_PROP(PhysicsScene, D6JointSwing2Motion), D6MotionEnum()),
-				enum_property("Twist", LUMIX_PROP(PhysicsScene, D6JointTwistMotion), D6MotionEnum()),
-				property("Linear limit", LUMIX_PROP(PhysicsScene, D6JointLinearLimit), MinAttribute(0)),
-				property("Swing limit", LUMIX_PROP(PhysicsScene, D6JointSwingLimit), RadiansAttribute()),
-				property("Twist limit", LUMIX_PROP(PhysicsScene, D6JointTwistLimit), RadiansAttribute()),
-				property("Damping", LUMIX_PROP(PhysicsScene, D6JointDamping)),
-				property("Stiffness", LUMIX_PROP(PhysicsScene, D6JointStiffness)),
-				property("Restitution", LUMIX_PROP(PhysicsScene, D6JointRestitution))
-			),
-			component("spherical_joint",
-				property("Connected body", LUMIX_PROP(PhysicsScene, JointConnectedBody)),
-				property("Axis position", LUMIX_PROP(PhysicsScene, JointAxisPosition)),
-				property("Axis direction", LUMIX_PROP(PhysicsScene, JointAxisDirection)),
-				property("Use limit", LUMIX_PROP(PhysicsScene, SphericalJointUseLimit)),
-				property("Limit", LUMIX_PROP(PhysicsScene, SphericalJointLimit), RadiansAttribute())
-			),
-			component("distance_joint",
-				property("Connected body", LUMIX_PROP(PhysicsScene, JointConnectedBody)),
-				property("Axis position", LUMIX_PROP(PhysicsScene, JointAxisPosition)),
-				property("Damping", LUMIX_PROP(PhysicsScene, DistanceJointDamping),	MinAttribute(0)),
-				property("Stiffness", LUMIX_PROP(PhysicsScene, DistanceJointStiffness), MinAttribute(0)),
-				property("Tolerance", LUMIX_PROP(PhysicsScene, DistanceJointTolerance), MinAttribute(0)),
-				property("Limits", LUMIX_PROP(PhysicsScene, DistanceJointLimits))
-			),
-			component("hinge_joint",
-				property("Connected body", LUMIX_PROP(PhysicsScene, JointConnectedBody)),
-				property("Damping", LUMIX_PROP(PhysicsScene, HingeJointDamping), MinAttribute(0)),
-				property("Stiffness", LUMIX_PROP(PhysicsScene, HingeJointStiffness), MinAttribute(0)),
-				property("Axis position", LUMIX_PROP(PhysicsScene, JointAxisPosition)),
-				property("Axis direction", LUMIX_PROP(PhysicsScene, JointAxisDirection)),
-				property("Use limit", LUMIX_PROP(PhysicsScene, HingeJointUseLimit)),
-				property("Limit", LUMIX_PROP(PhysicsScene, HingeJointLimit), RadiansAttribute())
-			),
-			component("physical_controller",
-				functions(
-					LUMIX_FUNC_EX(PhysicsScene::moveController, "move"),
-					LUMIX_FUNC_EX(PhysicsScene::isControllerCollisionDown, "isCollisionDown")
-				),
-				property("Radius", LUMIX_PROP(PhysicsScene, ControllerRadius)),
-				property("Height", LUMIX_PROP(PhysicsScene, ControllerHeight)),
-				property("Layer", LUMIX_PROP(PhysicsScene, ControllerLayer), LayerEnum()),
-				property("Use root motion", LUMIX_PROP(PhysicsScene, ControllerUseRootMotion)),
-				property("Use custom gravity", LUMIX_PROP(PhysicsScene, ControllerCustomGravity)),
-				property("Custom gravity acceleration", LUMIX_PROP(PhysicsScene, ControllerCustomGravityAcceleration))
-			),
-			component("rigid_actor",
-				functions(
-					LUMIX_FUNC_EX(PhysicsScene::putToSleep, "putToSleep"),
-					LUMIX_FUNC_EX(PhysicsScene::getActorSpeed, "getSpeed"),
-					LUMIX_FUNC_EX(PhysicsScene::getActorVelocity, "getVelocity"),
-					LUMIX_FUNC_EX(PhysicsScene::applyForceToActor, "applyForce"),
-					LUMIX_FUNC_EX(PhysicsScene::applyForceToActor, "applyImpulse"),
-					LUMIX_FUNC_EX(PhysicsScene::addForceAtPos, "addForceAtPos")
-				),
-				property("Layer", LUMIX_PROP(PhysicsScene, ActorLayer), LayerEnum()),
-				enum_property("Dynamic", LUMIX_PROP(PhysicsScene, DynamicType), DynamicTypeEnum()),
-				property("Trigger", LUMIX_PROP(PhysicsScene, IsTrigger)),
-				array("Box geometry", &PhysicsScene::getBoxGeometryCount, &PhysicsScene::addBoxGeometry, &PhysicsScene::removeBoxGeometry,
-					property("Size", LUMIX_PROP(PhysicsScene, BoxGeomHalfExtents)),
-					property("Position offset", LUMIX_PROP(PhysicsScene, BoxGeomOffsetPosition)),
-					property("Rotation offset", LUMIX_PROP(PhysicsScene, BoxGeomOffsetRotation), RadiansAttribute())),
-				array("Sphere geometry", &PhysicsScene::getSphereGeometryCount, &PhysicsScene::addSphereGeometry, &PhysicsScene::removeSphereGeometry,
-					property("Radius", LUMIX_PROP(PhysicsScene, SphereGeomRadius), MinAttribute(0)),
-					property("Position offset", LUMIX_PROP(PhysicsScene, SphereGeomOffsetPosition))),
-				property("Mesh", LUMIX_PROP(PhysicsScene, MeshGeomPath), ResourceAttribute("Mesh (*.msh)", PhysicsGeometry::TYPE))
-			),
-			component("wheel",
-				property("Radius", LUMIX_PROP(PhysicsScene, WheelRadius), MinAttribute(0)),
-				property("Width", LUMIX_PROP(PhysicsScene, WheelWidth), MinAttribute(0)),
-				property("Mass", LUMIX_PROP(PhysicsScene, WheelMass), MinAttribute(0)),
-				property("MOI", LUMIX_PROP(PhysicsScene, WheelMOI), MinAttribute(0)),
-				enum_property("Slot", LUMIX_PROP(PhysicsScene, WheelSlot), WheelSlotEnum())
-			),
-			component("physical_heightfield",
-				property("Layer", LUMIX_PROP(PhysicsScene, HeightfieldLayer), LayerEnum()),
-				property("Heightmap", LUMIX_PROP(PhysicsScene, HeightmapSource), ResourceAttribute("Image (*.raw)", Texture::TYPE)),
-				property("Y scale", LUMIX_PROP(PhysicsScene, HeightmapYScale), MinAttribute(0)),
-				property("XZ scale", LUMIX_PROP(PhysicsScene, HeightmapXZScale), MinAttribute(0))
-			)
-		);
-		registerScene(phy_scene);
-	}
-
-
 	struct CustomErrorCallback : physx::PxErrorCallback
 	{
 		void reportError(physx::PxErrorCode::Enum code, const char* message, const char* file, int line) override
@@ -248,6 +93,7 @@ namespace Lumix
 			, m_manager(*this, engine.getAllocator())
 			, m_physx_allocator(m_allocator)
 		{
+			PhysicsScene::reflect();
 			m_layers.count = 2;
 			memset(m_layers.names, 0, sizeof(m_layers.names));
 			for (u32 i = 0; i < lengthOf(m_layers.names); ++i)
@@ -259,7 +105,6 @@ namespace Lumix
 				m_layers.filter[i] = 0xffffFFFF;
 			}
 			
-			registerProperties(engine.getAllocator());
 			m_manager.create(PhysicsGeometry::TYPE, engine.getResourceManager());
 			LuaWrapper::createSystemFunction(engine.getState(), "Physics", "raycast", &LUA_raycast);
