+++ src/renderer/render_scene.cpp
@@ -17,6 +17,7 @@
 #include "engine/resource_manager.h"
 #include "engine/stream.h"
 #include "engine/universe.h"
+#include "imgui/IconsFontAwesome5.h"
 #include "renderer/culling_system.h"
 #include "renderer/font.h"
 #include "renderer/material.h"
@@ -2485,35 +2486,151 @@ private:
 	HashMap<Material*, EntityRef> m_material_decal_map;
 };
 
+void RenderScene::reflect() {
+	using namespace Reflection;
+
+	struct RotationModeEnum : Reflection::EnumAttribute {
+		u32 count(ComponentUID cmp) const override { return 2; }
+		const char* name(ComponentUID cmp, u32 idx) const override {
+			switch((Terrain::GrassType::RotationMode)idx) {
+				case Terrain::GrassType::RotationMode::ALL_RANDOM: return "All random";
+				case Terrain::GrassType::RotationMode::Y_UP: return "Y up";
+				default: ASSERT(false); return "N/A";
+			}
+		}
+	};
+
+	struct BoneEnum : Reflection::EnumAttribute {
+		u32 count(ComponentUID cmp) const override {
+			RenderScene* render_scene = static_cast<RenderScene*>(cmp.scene);
+			EntityPtr model_instance = getModelInstance(render_scene, (EntityRef)cmp.entity);
+			if (!model_instance.isValid()) return 0;
+
+			auto* model = render_scene->getModelInstanceModel((EntityRef)model_instance);
+			if (!model || !model->isReady()) return 0;
+
+			return model->getBoneCount();
+		}
+
+		const char* name(ComponentUID cmp, u32 idx) const override {
+			RenderScene* render_scene = static_cast<RenderScene*>(cmp.scene);
+			EntityPtr model_instance = getModelInstance(render_scene, (EntityRef)cmp.entity);
+			if (!model_instance.isValid()) return "";
+
+			auto* model = render_scene->getModelInstanceModel((EntityRef)model_instance);
+			if (!model) return "";
+
+			return idx < (u32)model->getBoneCount() ? model->getBone(idx).name.c_str() : "N/A";
+		}
 
 
-#define COMPONENT_TYPE(type, name) \
-	{ \
-		type \
-		, &RenderSceneImpl::create##name \
-		, &RenderSceneImpl::destroy##name \
-	}
+		EntityPtr getModelInstance(RenderScene* render_scene, EntityRef bone_attachment) const {
+			EntityPtr parent_entity = render_scene->getBoneAttachmentParent(bone_attachment);
+			if (!parent_entity.isValid()) return INVALID_ENTITY;
+			return render_scene->getUniverse().hasComponent((EntityRef)parent_entity, MODEL_INSTANCE_TYPE) ? parent_entity : INVALID_ENTITY;
+		}
+	};
 
-static struct
-{
-	ComponentType type;
-	void (RenderSceneImpl::*creator)(EntityRef);
-	void (RenderSceneImpl::*destroyer)(EntityRef);
-} COMPONENT_INFOS[] = {
-	COMPONENT_TYPE(MODEL_INSTANCE_TYPE, ModelInstance),
-	COMPONENT_TYPE(ENVIRONMENT_TYPE, Environment),
-	COMPONENT_TYPE(POINT_LIGHT_TYPE, PointLight),
-	COMPONENT_TYPE(DECAL_TYPE, Decal),
-	COMPONENT_TYPE(CAMERA_TYPE, Camera),
-	COMPONENT_TYPE(TERRAIN_TYPE, Terrain),
-	COMPONENT_TYPE(BONE_ATTACHMENT_TYPE, BoneAttachment),
-	COMPONENT_TYPE(ENVIRONMENT_PROBE_TYPE, EnvironmentProbe),
-	COMPONENT_TYPE(REFLECTION_PROBE_TYPE, ReflectionProbe),
-	COMPONENT_TYPE(PARTICLE_EMITTER_TYPE, ParticleEmitter),
-	COMPONENT_TYPE(FUR_TYPE, Fur)
-};
+	static auto render_scene = scene("renderer",
+		functions(
+			LUMIX_FUNC(RenderScene::setGlobalLODMultiplier),
+			LUMIX_FUNC(RenderScene::getGlobalLODMultiplier),
+			LUMIX_FUNC(RenderScene::addDebugCross),
+			LUMIX_FUNC(RenderScene::addDebugLine)
+		),
+		LUMIX_CMP(RenderSceneImpl, BoneAttachment, "bone_attachment", "Render / Bone attachment",
+			property("Parent", LUMIX_PROP(RenderScene, BoneAttachmentParent)),
+			property("Relative position", LUMIX_PROP(RenderScene, BoneAttachmentPosition)),
+			property("Relative rotation", LUMIX_PROP(RenderScene, BoneAttachmentRotation), 
+				RadiansAttribute()),
+			property("Bone", LUMIX_PROP(RenderScene, BoneAttachmentBone), BoneEnum()) 
+		),
+		LUMIX_CMP(RenderSceneImpl, Fur, "fur", "Render / Fur",
+			var_property("Layers", &RenderScene::getFur, &FurComponent::layers),
+			var_property("Scale", &RenderScene::getFur, &FurComponent::scale),
+			var_property("Gravity", &RenderScene::getFur, &FurComponent::gravity),
+			var_property("Enabled", &RenderScene::getFur, &FurComponent::enabled)
+		),
+		LUMIX_CMP(RenderSceneImpl, EnvironmentProbe, "environment_probe", "Render / Environment probe",
+			property("Enabled", &RenderScene::isEnvironmentProbeEnabled, &RenderScene::enableEnvironmentProbe),
+			var_property("Inner range", &RenderScene::getEnvironmentProbe, &EnvironmentProbe::inner_range),
+			var_property("Outer range", &RenderScene::getEnvironmentProbe, &EnvironmentProbe::outer_range)
+		),
+		LUMIX_CMP(RenderSceneImpl, ReflectionProbe,"reflection_probe", "Render / Reflection probe",
+			property("Enabled", &RenderScene::isReflectionProbeEnabled, &RenderScene::enableReflectionProbe),
+			var_property("size", &RenderScene::getReflectionProbe, &ReflectionProbe::size),
+			var_property("half_extents", &RenderScene::getReflectionProbe, &ReflectionProbe::half_extents)
+		),
+		LUMIX_CMP(RenderSceneImpl, ParticleEmitter, "particle_emitter", "Render / Particle emitter",
+			property("Emit rate", LUMIX_PROP(RenderScene, ParticleEmitterRate)),
+			property("Source", LUMIX_PROP(RenderScene, ParticleEmitterPath),
+				ResourceAttribute("Particle emitter (*.par)", ParticleEmitterResource::TYPE))
+		),
+		LUMIX_CMP(RenderSceneImpl, Camera, "camera", "Render / Camera",
+			var_property("FOV", &RenderScene::getCamera, &Camera::fov, RadiansAttribute()),
+			var_property("Near", &RenderScene::getCamera, &Camera::near, MinAttribute(0)),
+			var_property("Far", &RenderScene::getCamera, &Camera::far, MinAttribute(0)),
+			var_property("Orthographic", &RenderScene::getCamera, &Camera::is_ortho),
+			var_property("Orthographic size", &RenderScene::getCamera, &Camera::ortho_size, MinAttribute(0))
+		),
+		LUMIX_CMP(RenderSceneImpl, ModelInstance, "model_instance", "Render / Mesh",
+			functions(
+				LUMIX_FUNC_EX(RenderScene::getModelInstanceModel, "getModel")
+			),
+			property("Enabled", &RenderScene::isModelInstanceEnabled, &RenderScene::enableModelInstance),
+			property("Material", &RenderScene::getModelInstanceMaterialOverride,&RenderScene::setModelInstanceMaterialOverride, NoUIAttribute()),
+			property("Source", LUMIX_PROP(RenderScene, ModelInstancePath), ResourceAttribute("Mesh (*.msh)", Model::TYPE))
+		),
+		LUMIX_CMP(RenderSceneImpl, Environment, "environment", "Render / Environment",
+			var_property("Color", &RenderScene::getEnvironment, &Environment::diffuse_color, ColorAttribute()),
+			var_property("Intensity", &RenderScene::getEnvironment, &Environment::diffuse_intensity, MinAttribute(0)),
+			var_property("Indirect intensity", &RenderScene::getEnvironment, &Environment::indirect_intensity, MinAttribute(0)),
+			property("Shadow cascades", LUMIX_PROP(RenderScene, ShadowmapCascades)),
+			property("Cast shadows", LUMIX_PROP(RenderScene, EnvironmentCastShadows))
+		),
+		LUMIX_CMP(RenderSceneImpl, PointLight, "point_light", "Render / Point light",
+			property("Cast shadows", LUMIX_PROP(RenderScene, PointLightCastShadows)),
+			property("Dynamic", LUMIX_PROP(RenderScene, PointLightDynamic)),
+			var_property("Intensity", &RenderScene::getPointLight, &PointLight::intensity, MinAttribute(0)),
+			var_property("FOV", &RenderScene::getPointLight, &PointLight::fov, ClampAttribute(0, 360), RadiansAttribute()),
+			var_property("Attenuation", &RenderScene::getPointLight, &PointLight::attenuation_param, ClampAttribute(0, 100)),
+			var_property("Color", &RenderScene::getPointLight, &PointLight::color, ColorAttribute()),
+			property("Range", LUMIX_PROP(RenderScene, LightRange), MinAttribute(0))
+		),
+		LUMIX_CMP(RenderSceneImpl, Decal, "decal", "Render / Decal",
+			property("Material", LUMIX_PROP(RenderScene, DecalMaterialPath),
+				ResourceAttribute("Material (*.mat)", Material::TYPE)),
+			property("Half extents", LUMIX_PROP(RenderScene, DecalHalfExtents), 
+				MinAttribute(0))
+		),
+		LUMIX_CMP(RenderSceneImpl, Terrain, "terrain", "Render / Terrain",
+			functions(
+				LUMIX_FUNC(RenderScene::getTerrainNormalAt),
+				LUMIX_FUNC(RenderScene::getTerrainHeightAt)
+			),
+			property("Material", LUMIX_PROP(RenderScene, TerrainMaterialPath),
+				ResourceAttribute("Material (*.mat)", Material::TYPE)),
+			property("XZ scale", LUMIX_PROP(RenderScene, TerrainXZScale), 
+				MinAttribute(0)),
+			property("Height scale", LUMIX_PROP(RenderScene, TerrainYScale), 
+				MinAttribute(0)),
+			array("grass", &RenderScene::getGrassCount, &RenderScene::addGrass, &RenderScene::removeGrass,
+				property("Mesh", LUMIX_PROP(RenderScene, GrassPath),
+					ResourceAttribute("Mesh (*.msh)", Model::TYPE)),
+				property("Distance", LUMIX_PROP(RenderScene, GrassDistance),
+					MinAttribute(1)),
+				property("Density", LUMIX_PROP(RenderScene, GrassDensity)),
+				property("Mode", LUMIX_PROP(RenderScene, GrassRotationMode), RotationModeEnum())
+			)
+		)
+	);
+	registerScene(render_scene);
 
-#undef COMPONENT_TYPE
+	setIcon(ENVIRONMENT_TYPE, ICON_FA_GLOBE);
+	setIcon(POINT_LIGHT_TYPE, ICON_FA_LIGHTBULB);
+	setIcon(CAMERA_TYPE, ICON_FA_CAMERA);
+	setIcon(BONE_ATTACHMENT_TYPE, ICON_FA_BONE);
+}
 
 RenderSceneImpl::RenderSceneImpl(Renderer& renderer,
 	Engine& engine,
@@ -2554,10 +2671,12 @@ RenderSceneImpl::RenderSceneImpl(Renderer& renderer,
 	m_mesh_sort_data.reserve(5000);
 
 	m_render_cmps_mask = 0;
-	for (auto& i : COMPONENT_INFOS)
-	{
-		m_render_cmps_mask |= (u64)1 << i.type.index;
-		universe.registerComponentType(i.type, this, i.creator, i.destroyer);
+	
+	const u32 hash = crc32("renderer");
+	for (const Reflection::RegisteredComponent& cmp : Reflection::getComponents()) {
+		if (cmp.scene == hash) {
+			m_render_cmps_mask |= (u64)1 << cmp.cmp->component_type.index;
+		}
 	}
 }
